Update patch set 1
Patch Set 1:

The second of two patches to correct the problem with the TTS voice prompts being hard to hear when triggering VoiceDialer via Bluetooth.

In tonight's episode, we've switched Android's regular coffee with its seemingly arbitrary -18dB shift from the max volume level (which is about half-volume... far too quiet) with Folgers Crystals which rock a level that is about 75% of the max volume level.

Let's see if anyone notices.

Patch-set: 1